Trace calls using the pdata function index in the exe
Sometimes there is no return address on the stach when the breakpoints hit.
This might be due to some type of optimisation where functions are jumped to instead of called.
Or the pdata is wrong.
We could fix this by placing breakpoints on all rets in the function. Using Pdata to find the function boundary. This would require us to use code breakpoints which has a tendency to crash when running multithreded stuff.
Mabye we should look at DynamoRIO after all. Or build my own function injector that injects a jump to a function in the beginning of each function. But that might be complicated what if a function has ops that care about the instruction pointer location. Like relative jumps... DynamoRIO might be good after all...